Add SenderService to upload instances
