+++ share_app/src/main/java/org/odk/share/utilities/FileUtils.java
@@ -1,0 +1,18 @@
+package org.odk.share.utilities;
+
+import java.util.Locale;
+
+public final class FileUtils {
+
+    private FileUtils() {
+
+    }
+
+    public static String getFileExtension(String fileName) {
+        int dotIndex = fileName.lastIndexOf('.');
+        if (dotIndex == -1) {
+            return "";
+        }
+        return fileName.substring(dotIndex + 1).toLowerCase(Locale.ROOT);
+    }
+}
