+++ share_app/src/main/res/values/strings.xml
@@ -52,5 +52,6 @@
     <string name="receive_success">Successfully received %1$s form(s)</string>
     <string name="send_success">Successfully sent %1$s form(s)</string>
     <string name="error_while_downloading">Error occurred while downloading form</string>
-    <string name="download_queued">Downloading queued...</string>
+    <string name="download_queued">Downloading queued…</string>
+    <string name="upload_queued">Uploading queued…</string>
 </resources>
